Reserch/requirements.txt
# Google API Client Libraries
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Authentication
google-auth==2.27.0